requirements.txt
# ========== Framework Web ==========
fastapi==0.104.1          # Framework web moderno y rápido para APIs
uvicorn==0.24.0           # Servidor ASGI para ejecutar FastAPI

# ========== Base de Datos ==========
sqlalchemy>=2.0.36        # ORM para manejo de base de datos
psycopg2-binary==2.9.11   # Driver de PostgreSQL para Python

# ========== Autenticación y Seguridad ==========
python-jose==3.3.0        # Manejo de tokens JWT
passlib==1.7.4            # Hashing de contraseñas
bcrypt==4.0.1             # Algoritmo de encriptación para contraseñas

# ========== Procesamiento de Archivos ==========
python-multipart==0.0.6   # Manejo de uploads de archivos
pandas>=2.1.0             # Procesamiento de datos Excel
openpyxl>=3.1.2           # Lectura/escritura de archivos Excel (.xlsx)

# ========== Templates y Frontend ==========
jinja2==3.1.2             # Motor de templates HTML

# ========== Configuración ==========
python-dotenv==1.0.0      # Manejo de variables de entorno desde .env
